+++ docs/releases.md
@@ -17,6 +17,7 @@ changes over time.
 ### Version 0.26.1
 
 * github-226: Upgrade to Spark 3.2.2
+* github-227: [BUG] Flowman should not fail with field names containing "-", "/" etc
 
 
 ### Version 0.26.0 - 2022-07-27